Track allocation of non-preemptible tasks separately
Summary:
To support the notion of non-preemptible tasks, the resource manager needs to
start tracking their allocation separately. This is required for admission control of
non-preemptible tasks in a resource pool. This diff only adds the ability to track
the allocation of  non-preemptible tasks, it doesn't affect admission control
or preemption. Those will be addressed in later diffs.

Reviewers: #peloton, mabansal, apoorvaj

Reviewed By: #peloton, apoorvaj

Subscribers: min, mu, jenkins

Maniphest Tasks: T1534303

Differential Revision: https://code.uberinternal.com/D1485573