Evaluate and update job state even if the task stats have not changed.
Summary:
Due to previous bugs, there are multiple jobs stuck in production with
task stats in job runtime are accurate, but the job state stored is
incorrect. Adding code to evaluate and update job even if the task
stats have not changed to recover those jobs.

Resolves T1523235 T1519675

Reviewers: mu, min, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins

Maniphest Tasks: T1523235, T1519675

Differential Revision: https://code.uberinternal.com/D1494015