Fix demand calculation in entitlement calculator
Summary:
For reliable floating point calculations,  the entitlement calculation uses `ResourceEpsilon = 0.009` as the minimum
for scalar resource representation. This diff fixes a bug where demand for a
resource type would return true if it was greater than 0. This diff changes the
truthiness of the method `demandExists` to return true only if the demand is
greater than `ResourceEpsilon`.

There could be cases where the demand is greater than zero but less than ResourceEpsilon
because of which the entitlement calculation can go into an infinite loop.

Reviewers: apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T1548185

Differential Revision: https://code.uberinternal.com/D1499951