Add API to be able to fetch different task (runs) of an instance
Summary:
Two API changes are being added.
1. Modify task get to fetch the last runtime of each task of an instance instead of just the last one. This runtime will have the mesos-task-id, hostname and the mesos agentid.
2. Modify task browse sandbox to fetch the path to the logs of an older task of an instance. It takes the mesos-task-id (as a string) as an input if a user wants to fetch the path to the logs of an older task.

So, the workflow is that a user first uses task get to fetch the list of current task as well as previous tasks. Each task runtime will have the mesos task id of the current task run as well as the mesos task id of the previous task run. The user will then use the mesos-task-id of the previous runs in the browse sandbox to fetch the path to the logs of an older task.

Reviewers: min, mu, jpinkul, #peloton

Reviewed By: min, #peloton

Subscribers: zhitao, avyas, jenkins

Differential Revision: https://code.uberinternal.com/D1513083