Add support for fetching task logs for previous task runs
Summary:
An instance may be run multiple times, but currently there is no way to
fetch the logs from an older task run. Adding handler for two APIs.
The first is a task get API which returns the terminal state of
all the task runs of an instance. It will return the final state,
the mesos-task ID and the previous mesos task ID of each task run.
Then the user can invoke the BrowseSandbox API to fetch the logs
of any task run by giving a particular mesos task ID of a
particular task run as an input.

Resolves T1565431

Test Plan:
1. Tested on local pcluster.
2. Unit tests pass.

Reviewers: min, mu, avyas, adityacb, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins, jpinkul

Maniphest Tasks: T1565431

Differential Revision: https://code.uberinternal.com/D1521549