Fix sorting in job query response
Summary:
There was a bug in QueryJobs, that ignored sorting if there was
no keyword or search clause specified. So sort worked if a keyword
was specified but not otherwise.

Limit was in essence made to be compulsory when specifying pagination,
and hence if limit was not set, we defaulted limit to 0 and
returned empty results. Fixed that and now we default to limit 10.

Added unit tests.

Peloton CLI now will look for JobInfo as well as JobSummary as
long as we are returning jobInfo in the query result.

Fixed backend sorting bug for T1122619

Reviewers: #peloton, apoorvaj, min

Reviewed By: #peloton, apoorvaj

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1515923