[API] Add Controller task type in the peloton API
Summary:
This diff adds a new task type `CONTROLLER` and a new field in the
`TaskConfig` to define if a task is controller via a boolean `controller`.
This is required to differentiate the controller task inside peloton
for supporting spark use cases. The user can specify if a task is a
controller task by flipping the `controller` boolean in the task config.
By default it is false.

An example use case for using controller tasks would be to support the spark
driver and executors in the same job. The spark driver in this case would
be marked as the controller task.

Reviewers: apoorvaj, min, jpinkul, #peloton

Reviewed By: min, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1520779