Persist FAILED task state into DB even if task needs to be restarted
Summary:
Currently, if a task fails, and it needs to be restarted, the task state
is moved directly to INITIALIZED and the terminal state FAILED is not
persisted into the DB. This causes a user to see a task move from
RUNNING state to INITIALIZED with the message indicating that it failed.
This will cause issue when we support a customer being able to see
multiple task runs of the same instance as a customer will not see
a terminal state for an older task run.

With this change, the FAILED task state is persisted into DB when
FAILED state is received from mesos. And goal state is used to restart
the task after failure.

Resolves T1569775

Reviewers: mu, min, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins

Maniphest Tasks: T1569775

Differential Revision: https://code.uberinternal.com/D1525255