API: Add summaryOnly flag to QueryRequest
Summary:
Currently, JobQuery response contains []*JobInfo as well
as []*JobSummary. JobInfo is entire jobconfig and is a high
latency operation involving a lucene query to get job_ids
followed by a DB query to get job config for the job_ids.

JobSummary is just few fields from job_index that we
retrieve by lucene query only and is quicker and more
light weight in terms of data transfer.

For the time being we will continue to support JobInfo
and JobSummary as part of query response. But we will
support a "summaryOnly" flag in the Query Request
that will tell the handler to just return JobSummary
as party of QueryResponse, and not construct JobInfo at all.
This will help clients like Peloton UI in the short term
to speed up their JobQuery response.

Resolves T1573911

Test Plan: Added unit tests for this use case.

Reviewers: #peloton, min, apoorvaj

Reviewed By: #peloton, min

Subscribers: jenkins

Maniphest Tasks: T1573911

Differential Revision: https://code.uberinternal.com/D1526057