[API] Add controller limit in the resource pool config
Summary:
This diff adds a `ControllerLimit` configuration in the resource pool
config. If set this defines a maxmimum limit of resources `CONTROLLER` tasks
can use in a resource pool. It is calculated as a percentage of the resource
pool's reservation. If unset controller tasks can use all of the entitled resoruces.

Reviewers: apoorvaj, min, jpinkul, #peloton

Reviewed By: min, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1527889