JobQuery optimization
Summary:
With queries that end up matching a huge number of rows
from job_index table, sorting takes a long time to complete
and LIMIT is applied on the sorted result.
To shrink the result set to a manageable number, we will
search for jobs that have been created within past 7 days
only. This will be made customizable in a future diff.
I have tested the new query on pit1_prod01 and pit1_preprod01

Test Plan: Added unit test.

Reviewers: #peloton, apoorvaj

Reviewed By: #peloton, apoorvaj

Subscribers: apoorvaj, jenkins

Differential Revision: https://code.uberinternal.com/D1536959