Terminate the statemachine when a task is removed from the tracker
Summary:
When a task is removed from the tracker the task statemachine
should terminate and perform cleanup eg. stop the state timer if the task
is in a temout state. Otherwise the callbacks will fail because the task is
no longer there in the tracker.

Test Plan: Added unit tests

Reviewers: #peloton, apoorvaj, mu, min

Reviewed By: #peloton, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T1558243

Differential Revision: https://code.uberinternal.com/D1542103