Change controller task interface between jobmgr and resmgr
Summary:
This reverts the use of a new `TaskType` to differentiate between controller and non-controller tasks.
Instead this changes it by using a boolean controller flag instead in the TaskConfig being sent to the resource
manager from the job manager. This also makes the interface between the resource manager and the placement
engine much cleaner by not exposing internals a task type.

This change removes the prior controller `TaskType` from the protobuf instead of deprecating it because:
1. job manager and resource manager are deployed together.
2. no one is using this feature yet.

Reviewers: #peloton, apoorvaj, min, mabansal

Reviewed By: #peloton, apoorvaj

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1545611