Add a common library to implement a goal state engine
Summary:
The goal state engine has two parts.
The first is a deadline queue implementation which provides
an Enqueue function to enqueue jobs into the deadline queue
and a blocking Dequeue operation to wait for an enqueue item
whose deadline expires.
The second is a goal state engine implementation which implements
an Enqueue function which allows the caller to enqueue an entity
into the goal state engine. After its deadline expires, the engine
callbacks to fetch the state and the goal state of the entity,
find the action list to executed for the given state and goal state,
and executes each action in the action list. If any of the actions
return error, the entire list is executed again with an expoential
backoff till the actions succeed.

Resolves T1596515

Test Plan: Adding unit tests for both the modules.

Reviewers: min, mu, mabansal, #peloton, pourchet

Reviewed By: mu, #peloton

Subscribers: jenkins

Maniphest Tasks: T1596515

Differential Revision: https://code.uberinternal.com/D1592543