Remove 7 day time span restriction from querying active jobs
Summary:
In 0.6.10.3, currently we put a restriction on job query to
fetch jobs that were created 7 days ago. This was done to improve
performance by restricting search criteria.

However, ATG might have jobs that are active and have been created
before that 1 week span and these won't get displayed.

Long term fix for this is to treat job query for active and completed
jobs differently in UI and restrict completed job search to last
7 days and not restricy active job search to any time span by default.
Also, we are working to add job query by custom time duration
(last n days) to make this more customizable.

This is a temporary fix to mitigate the problem in ATG where
a user who has active jobs running for more than a week should
still be able to see them in peloton UI.

Reviewers: #peloton, apoorvaj, min, mabansal

Reviewed By: #peloton, mabansal

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1554077