Fix race condition in state machine rollback
Summary:
This diff tightens the scope of the lock on the statemachine when recovery
is performed from a timeout state. The lock should be held for the duration of the
statemachine rollback and the termination of the state timer.

This diff also changes the behaviour when a statemachine cannot be terminated cleanly by
logging and the error and continuing.

Test Plan: ran tests locally

Reviewers: mabansal, #peloton

Reviewed By: mabansal, #peloton

Subscribers: jenkins

Maniphest Tasks: T1606205

Differential Revision: https://code.uberinternal.com/D1558373