Correct scheduled task accounting in case of launch errors for maxRunningInstance feature
Summary:
For the maximum running instances feature, accurate accounting of tasks
in running/pending/launched state is needed. Inaccuracies were introduced
due to the following.
1. If tasks enter initialized state from any other state, they were being sent
to resource manager directly instead of the job SLA evaluation routine.
This is fixed in this code.
2. If the mesos-task-id of a task changes, it is sent again to resource
manager, but resource manager sends back an error. Thus job manager
believes the task to be not present in resource manager even though it
does exist. This leads to a task remaining in INITIALIZED state even though
resource manager has the task. Fixing to have the resource manager send back
no error in case mesos-task-id of an existing task changes.

Also fixed the logic to handle the conditions when the number of scheduled
instances become larger than the maximum running instances.
Added logs and a metric to detect this condition as well.

Resolves T1626201

Test Plan:
1. Added unit tests.
2. Simulated these errors on local pcluster and ensured that peloton
behaves correctly.

Reviewers: mu, min, #peloton, mabansal

Reviewed By: mu, #peloton

Subscribers: jenkins

Maniphest Tasks: T1626201

Differential Revision: https://code.uberinternal.com/D1582195