Do not update the state transition reason on dequeue from placement engine
Summary:
Make the reason to be passed during state transition variadic allowing
the caller to decide whether to update the previous transition reason or
retain the previous reason. It is currently used to not update the
transition reason when the task is dequeued for placement.
This allows retaining the reason when the task is returned from placement.
With this change, the reason why task did not get placed for 10mins
shows up in the state machine reason.

bin/peloton resmgr tasks active
                                  TaskID|    State|                                                          Reason|Last Update Time
  737ab358-2ff3-4890-944a-4971ca64a013-0|  PLACING|  Previous placement timed out due to {"mismatch_constraints":3}|2018-03-20 18:45:55.9466999 +0000 UTC

Resolves T1630507

Reviewers: mabansal, avyas, #peloton

Reviewed By: avyas, #peloton

Subscribers: jenkins

Maniphest Tasks: T1630507

Differential Revision: https://code.uberinternal.com/D1583145