Adding errorcodes in communication between resmgr and jobmgr for enqueuegangs
Summary:
Right now there are no errrocodes in communication between jobmgr and resmgr
as a result of it , jobmgr can not distinguish between different error cases
when it tries to submit tasks in resmgr. This leads to jobmgr to submit same task
again and again to ensure it will eventually succeed after getting the confirmation
from hostmgr. This leads to lot of unnecessary rpc's between job mgr and resmgr.
This also leads to have issues in exactly knowing how many tasks are running in remgr
to correctly detrmine the maxrunninginstances at a single point.
This diff addresses the errocodes in resmgr.

Test Plan: Tests added , Pcluster testing

Reviewers: #peloton, #protobuf_reviewers, avyas

Reviewed By: #peloton, avyas

Subscribers: pedge, jenkins

Differential Revision: https://code.uberinternal.com/D1620517