Bump up C* timeouts and add timers to recovery code
Summary:
During 04-05 outage, we saw that on rebooting jobmgr and
resmgr, the recovery code failed because of cassandra timeouts.
At that time we hot fixed this issue by bumping up cassandra
timeouts to 20sec for all components. Pulling that fix here.

We also observed that GetJobsByStates would need to be timed
to figure out why C* timeouts happen and for which state.
Added those metrics.

We also observed in case of one job_id that after deleting
job from job_runtime, the job still remains as part of
the materialized view queried by GetJobsByStates call.
This resulted in recovery code to bail out because it
could not find the job in job_runtime table. Instead of
bailing out, added code to log this and still continue with
recovery. Added unit test for this code.

Reviewers: #peloton, mabansal, min

Reviewed By: #peloton, mabansal

Subscribers: mu, avyas, jenkins

Differential Revision: https://code.uberinternal.com/D1646079