Add hostmgr recovery to restore the contents of maintenance queue
Summary:
When host maintenance is triggered, the hosts to be drained are enqueued in the maintenanceQueue. The resmgr polls this queue from time to time to determine which tasks need to be rescheduled. But if hostmgr crashes before the resmgr polls this queue then tasks placed on these hosts will not be rescheduled. This diff adds recovery code for hostmgr so that on hostmgr crash / leader change, the contents of maintenanceQueue are restored.

On recovery, the maintenanceQueue contents are restored as follows
1. Read the current Maintenance Status from the Mesos Master
2. Get machines in 'DRAINING' state from the Maintenance Status
3. Enqueue, into maintenanceQueue, the hostnames of machines in 'DRAINING' state

Test Plan: Unit tests

Reviewers: #peloton, varung

Reviewed By: #peloton, varung

Subscribers: rcharles, varung, jenkins

Differential Revision: https://code.uberinternal.com/D1687209