Deleted tasks from the gangs needs to be removed from gang and enqueued again
Summary:
Currently resmgr scheduler drops the gang on floor if there are tasks in the gang which
are invalid or could not be transitioned to correct state. This diff removes the invalid
tasks from the gang and rest of the gang is been pushed back to pending queue after transition.
This diff also addresses multiple sentry errors which should not ber error rather Info

Test Plan:
Unit Tests included

Pcluster Testing

1. ran jobs with pcluster
2. killed the tasks in job
3. all the non killed tasks succeeded

Reviewers: mu, avyas, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1723145