Fix flaky unit test in TestBucketEventProcessor
Summary:
TestBucketEventProcessor is flaky because it depends on time.Sleep() to control the
test flow. It is suspectible to scheduling and changes in other part of the code.

Also bucket.shutdown() would just send a shutdown signal and return without
ensuring goroutine actually ends, which results in data race in unit test.

The diff fixes these two scenarios by making sure when shutdown() is called,
it would wait for the goroutine to exit and all the existing events to finish.
Since shutdown() is used only in test, this diff should have no effect in production.

Reviewers: #peloton, apoorvaj, mu, varung

Reviewed By: #peloton, varung

Subscribers: jenkins

Maniphest Tasks: T1763145, T1749583

Differential Revision: https://code.uberinternal.com/D1733333