Adding GetHosts api from hostmanager to returns the list of hosts which matches the resources and constraints passed by placement engine
Summary:
Adding an internal API GetHosts from Host Manager to returns the list of hosts
which matches the resources as well as constraints/atrributes on the host.
As of now it happens only for Offers from Mesos, However we need it for hosts too.
There are two reasons we need it for
  - To make the Host reservation from the placement engine
  - To get to better placement strategies in the future.
Once Placement engine will get the hosts list which mathes the filter, it will
try to put the host reservation in Host Manager.

Test Plan:
unit tests added

Pcluster Testing
Changed temporary placement engine code and tested through placement engine to get the Hosts list

Reviewers: mu, avyas, #peloton, varung

Reviewed By: mu, #peloton

Subscribers: jenkins, varung

Differential Revision: https://code.uberinternal.com/D1752277