Reduce number of EnqueueJob calls
Summary:
Porting this change to 0.6.13.
This commit has the following fixes.

Reduce the number of EnqueueJob calls by calling one EnqueueJob for a batch of task updates instead of calling it once per task update.
Increase number of job worker threads in goalstate to 50 from 10.
Remove get active task calls for task list and task query till the deadlock in resource manager is fixed. Also, this change is impacting resource manager throughput as it acquires a global tracker lock. Hence, this change needs to be revisited.

Reviewers: zhixin, #peloton

Reviewed By: zhixin, #peloton

Differential Revision: https://code.uberinternal.com/D1763787