Fix deadlock in task tracker
Summary:
This diff fixes a deadlock in the task tracker which can occur in the following condition:
1. RMTask state machine rollback takes a lock on the statemachine and the callback takes a lock on the tracker
2. The `GetActiveTasks` API takes a lock on the tracker and then on the statemachine.

If `GetActiveTasks` is called quite frequently it can trigger the condition for a particular RMTask which would lock
the tracker completely.

This diff changes the mutex on the tracker to a `RWMutex` which is more granular, so in the case described above
the lock to be acquired on the tracker would be a `RLock` in 1 and 2 which should succeed.

Reviewers: #peloton, mabansal, apoorvaj

Reviewed By: #peloton, mabansal

Subscribers: jenkins

Maniphest Tasks: T1787107

Differential Revision: https://code.uberinternal.com/D1791831