Partially created job set to INITIALIZED and enqueue to goalstate engine
Summary:
Partially created job should be set to INITIALIZED and send to goalsate
engine, so goalstate engine can create the missing tasks for the job.
Previously, partially created job would not be set to INITIALIZED, which
results in a job stuck at PENDING/KILLING state. Also, previously,
partially created job would not be enqueued into goalstate engine even if
the state is set correctly. As a result, parially created job which all the
tasks enter terminal state would not be evaluated and stuck at the state.

The diff fixes the bug, and breaks down the unit test which was too big
to modify.

Reviewers: apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T1791681, T1701129

Differential Revision: https://code.uberinternal.com/D1795801