Add support for upgrading a task in the task goal state engine
Summary:
An upgrade is trigerred when the desired configuration version is different
from the the configuration version. If the task is running, then kill it.
Once it is reaches terminal state, then initialize it again.
On re-initialization, bring the task up with the desired configuration version.
If a task has terminal goal state, then ignore the upgrade request.

Resolves T1833349

Reviewers: zhixin, mu, #peloton

Reviewed By: zhixin, #peloton

Subscribers: jenkins

Maniphest Tasks: T1833349

Differential Revision: https://code.uberinternal.com/D1860247