Adding Reserver in placement engine for host reservation
Summary:
This diff is adding Reserver in Placement Engine.
Reserver is the async daemon which is running always and consuming tasks
from the placement engine and try to reserve the hosts.

Placement Engine and Reserver communicate using common/queue which is
internally implemented as channel.

This diff also adds the new Model/Host which is the mapping of HostInfo
from Host Manager as well as running tasks from resource manager.

Reserver gets the task from Placement engine and based on the filters
it tries to gethost and then after match try to reserve host.
Reserver follows below steps
   1. Get Tasks from the reservation queue
   2. Find the hosts list from hostmanager matching filter
   3. choose one random host from the list
   4. reserve the host in host manager

This diff also changing the model/host -> model/host offers, as this is what
it should be as this modeling host offers from HostManager

This diff is also adding hosts/Service which is calling hostmanager.GetHosts as
well as hostmanager.ReserverHosts

Test Plan:
 - Added New Tests
 - Modified Old Tests
 - Pcluster Tests for running Local Cluster

Reviewers: mu, #peloton, varung

Reviewed By: #peloton, varung

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1812479