Changing resource manager api from MarkTasksLaunched to UpdateTasksState for updating any state into Resourcemanager
Summary:
Currently we have MarkTasksLaunched api resource manager which makes list of tasks specified to LAUNCHED state
This diff is modifying this API to be more generic UpdateTasksState which will take a map of peloton
task to state map. This api will take the appropriate action based on the sate of the task.

UpdateTasksState will be called to notify the resource manager about the tasks
which have been moved to cooresponding state , by that resource manager
can take appropriate actions for those tasks. As an example if the tasks been
launched then job manager will call resource manager to notify it is launched
by that resource manager can stop timer for launching state. Similarly if
task is been failed to be launched in host manager due to valid failure then
job manager will tell resource manager about the task to be killed by that
can be removed from resource manager and relevant resource accounting can be done.

Test Plan: Tests Added, PCluster

Reviewers: mu, avyas, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1826149