Clean job upon create failure only if job does not exist before
Summary:
Now the Create handler would always try to clean up the job when it fails
to create the job in DB. This is intented to clean up DB, as it is not
certain if the job is already persisted in DB.
This should not include AlreadyExist err. Otherwise, the running job would
be removed.

Reviewers: #peloton, mu, apoorvaj

Reviewed By: #peloton, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T1861713

Differential Revision: https://code.uberinternal.com/D1855137