Handle job runtime update when more instances running than expected
Summary:
There was a race condition bug, which caused more tasks are running than
configured. This is not expected in JobRuntimeUpdater, and the func could
not correctly handle this situation. As a result, such jobs are stuck in
jobmgr.

This diff log an err if the bug is hit again. Also JobRuntimeUpdater handles
this case correctly. It will unblock any stuck job now, and prevent more
jobs to be stuck for similar bug.

Reviewers: mu, apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T1868231

Differential Revision: https://code.uberinternal.com/D1857119