[API] Add resourceUsage map to RuntimeInfo proto
Summary:
We need to add vCoresSeconds for CPU and GPU and vMemorySeconds into RuntimeInfo
for job as well as task. This diff adds these to the API proto message so that
clients can start integrating with this.

These will be added as a map<string>double where the key will be
resource kind i.e. "cpu", "gpu" and "memory" for now. Can be easily
extended to other kinds.

Reviewers: #peloton, min

Reviewed By: #peloton, min

Subscribers: min, rcharles, jenkins

Differential Revision: https://code.uberinternal.com/D1876397