Controller job decides terminal state with controller task state
Summary:
If a job has controller task, its terminal state should be determined
by terminal state of controller task. The code assumes that if there
is controller task in a job, there will be only one task and it will
be task[0].
A separate diff would be sent out to enforce this assumption during
job config validation.

Reviewers: mu, apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: avyas, jenkins

Maniphest Tasks: T1706139

Differential Revision: https://code.uberinternal.com/D1894659