Fix nil panic when update cache is not found
Summary:
In golang, an interface value is nil only if both type and value are nil.
If we directly return a value from map, chances are the type is non-nil and
value is nil. As a result, when using == to compare with nil, == will give
false.
Need to return nil directly to make == work when comparing interface with nil.

Reviewers: mu, apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T1919215

Differential Revision: https://code.uberinternal.com/D1913223