Recover Update upon JobMgr restart
Summary:
When JobMgr restarts and recover jobs, check if the job has update id.
If yes, enqueue update to goal state engine and rely on it to recover
cache.

Upgrade on terminated job is permitted, so JobMgr would also recover
terminated job which has an active update. As a result, Update removes
terminated job from cache once it is done.

This diff is based on D1908823

Test Plan:
in pcluster, make an update on a job. Stop jobmgr leader container. Restart
the leader. And update can still progress.

Reviewers: mu, apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T1780771

Differential Revision: https://code.uberinternal.com/D1912733