Implement update with batch size
Summary:
Update respects batch size in update config now. The diff removes
upgrade logic in UpgradeStart. All of the upgrade happens in
UpgradeRun.

UpgradeRun processes upgrade in the following order:
1. Check the current upgrade process
2. Calculate which instances to update in this run
3. Process upgrade
4. Update update progress

This diff does not enable add instances in update, and this
would be handled in T1925337.
Also the diff does not address update KILLED task, and this
issue would be address in T1926359.

Test Plan:
Create a job with twenty tasks.
Update job with batch size 1.
Update is processed 1 by 1.
All the tasks are updated (by checking the output of task,
which is different across the versions).

Reviewers: mu, apoorvaj, #peloton

Reviewed By: mu, #peloton

Subscribers: sachins, jenkins

Maniphest Tasks: T1774751

Differential Revision: https://code.uberinternal.com/D1921169