Enqueue initialized tasks being updated into goal state engine
Summary:
For task being updated, it can be in INITIALIZED state (e.g.
task is created in update but fails in db write and is not sent
to ResMgr). JobMgr need to enqueue such tasks into goal state engine,
otherwise those tasks would be stuck and cannot move on.

This diff is based on D1930483

Test Plan:
tested on pcluster by intentionally creating an error when create
tasks in update.

Reviewers: mu, apoorvaj, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins

Maniphest Tasks: T1931551

Differential Revision: https://code.uberinternal.com/D1932597