Merge instance config when update config
Summary:
In batch job update, instancesConfig of newConfig only need to include the config
for the additional instances. If a job config is directly updated to newConfig,
JobMgr would lose the track of previous instance config. As a result, JobMgr has
to use the merged result of instanceConfig in oldConfig and newConfig.
configs passed in mergeInstanceConfig must have been validated.

It solves the issue of a job cannot identify it contains a controller task after update.
An example of the failure:
A job is created with one controller task, which is configured by instance config.
The job updates its config with more instances, and only provides the new instance configs.
Job config does not have the instance config with the controller task config.
Job does not know it has a controller task anymore.

Test Plan:
In plcuster, create a job with one controller task which exit successfully.
Then update the job with more instances.
Kill instances in the job.
job status is SUCCESS

Reviewers: mu, apoorvaj, mabansal, #peloton, adityacb

Reviewed By: mu, #peloton, adityacb

Subscribers: min, avyas, adityacb, jenkins

Maniphest Tasks: T1933103

Differential Revision: https://code.uberinternal.com/D1930901