Reenqueue update if any task updated in UpdateRun is KILLED
Summary:
The existing code evaluate update upon task event.
But for KILLED task, there would be no task event sent after an update.
As a result, update would need to wait for other tasks updated to send
an event to carry on an update, which results in unnecessary delay.
Also if all of the tasks are KILLED, update could not make any progress.
Therefore, it any task is in KILLED state in the update batch, reevaluate
update again to resume update immediately.

This diff is based on D1921169

Reviewers: mu, apoorvaj, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins

Maniphest Tasks: T1926359

Differential Revision: https://code.uberinternal.com/D1930483