Add job configuration validation for different type of job
Summary:
Ref T1764969
This diff add functions in jobmgr/job/config.validateTaskConfigWithRange() to check on some specific fields that is different for batch job and stateless job. The function is called when job create and job update, so that the validation will happen meantime. This is to make sure that batch configuration is not specified for long running jobs, and long running services configuration is not provided for batch jobs

Reviewers: apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: min, mabansal, varung, jenkins, zhixin

Tags: #peloton

Maniphest Tasks: T1764969

Differential Revision: https://code.uberinternal.com/D1945917