Set task goal state to KILLED for terminated task
Summary:
As of now, a task in terminated state does not mean the task is
indeed terminated. For batch task, a failed task can be retried.
For stateless tasks, a task in SUCCEEDED/FAILED state is retried
as well.
If a task kill ignore tasks in terminated state, these task would
be running again when retry kicks in, which is unexpected to user.

Reviewers: mu, apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T1901857

Differential Revision: https://code.uberinternal.com/D1933245