JobRuntimeUpdater treats job with update as non-paritally-created job
Summary:
Now, if a job has an update which adds more instances, it would be
treated as partially created job and set to INITIALIZED state.
As a result, all of the remaining tasks would be created without
respecing update batch size,
Since, JobMgr updatesvc handler ensures an update would not happen on
partially created job, it is safe to assume that if a job has an update
and its total number of instances is smaller than configured, it is due
to update adds more instances.
JobRuntimeUpdater should differentiate a job with update from job without
update.

Test Plan: tested with pcluster

Reviewers: mu, apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T1941999

Differential Revision: https://code.uberinternal.com/D1941123