Enable archiver streaming only mode, add unit tests
Summary:
After talking to filebeat team, it seems that we can enable logging
"completed_job" event to stdout of archiver and they can configure
filebeat to stream archiver logs to logstash. They can then configure
logstash to catch this event and stream it via a kafka heatpipe topic
to HIVE. This fits our usecase and is worth trying till we figure out
if we can utilize cassandra CDC (end of Q4)

Adding archiver stream-only mode to test this out. If this mode is
enabled, then archiver will only log jobs to stdout. This will be
streamed to logstash by filebeat. Filebeat team will configure
logstash to send this to a heatpipe kafka topic (schema created by
peloton team)

I will run this on dca1-prod02 or sjc1-prod02 and try to setup
the hive pipeline and verify the data on Hive. If this works, then
we can decide whether to wait for CDC or go ahead with this.

Added/Improved tests, bump up coverage above 90%

Resolves T1976103, T1976113

Reviewers: #peloton, apoorvaj

Reviewed By: #peloton, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T1976113, T1976103

Differential Revision: https://code.uberinternal.com/D1980831